requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
python-telegram-bot
python-telegram-bot-calendar
fastnumbers
